# swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "version":"1.0.0",
        "description":"/API for PostAway application",
        "title":"PostAway App"
    },
    "servers":[
        {
            "url":"http://localhost:3000"
        }
    ],
    "components":{
        "securitySchemes":{
            "JWT":{
                "type":"/apiKey",
                "in":"cookie",
                "name":"jwtToken"
            }
        },
...
    },
    "paths":{
...
        "/api/otp/send": {
            "post": {
              "summary": "Send OTP to user",
              "description": "Generates and sends an OTP to the user's email address. Requires authentication via JWT.",
              "operationId": "sendOTP",
              "tags": ["OTP"],
              "security": [
                {
                  "JWT": []
                }
              ],
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "format": "email",
                          "description": "The email address where the OTP will be sent."
                        }
                      },
                      "required": ["email"]
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "OTP sent successfully",
                  "content": {
                    "text/plain": {
                      "schema": {
                        "type": "string"
                      }
                    }
                  }
                },
                "400": {
                  "description": "Internal error",
                  "content": {
                    "text/plain": {
                      "schema": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
        },
...
    }  
}